+++ app/renderer/epics/test.epics.js
@@ -1,8 +1,10 @@
 import { EMPTY, from, of } from 'rxjs';
-import { mergeMap } from 'rxjs/operators';
+import { concat, mergeMap } from 'rxjs/operators';
 import { ofType } from 'redux-observable';
+import { Wallet } from "cennznet-wallet";
 
 import types from '../types';
+import { storageKeys } from '../api/utils/storage';
 
 const testPageEpic = action$ =>
   action$.pipe(
@@ -10,4 +12,18 @@ const testPageEpic = action$ =>
     mergeMap(() => EMPTY)
   );
 
-export default [testPageEpic];
+const createWalletEpic = action$ =>
+  action$.pipe(
+    ofType(types.walletCreate.triggered),
+    mergeMap(async () => {
+      console.log('createWalletEpic');
+      const mnemonic = window.odin.api.cennz.createMnemonic({num: 24});
+      const wallet = await window.odin.api.cennz.createWallet({ name: 'test wallet', mnemonic, passphrase: 'password'});
+      return {
+        type: types.setStorage.requested,
+        payload: { key: storageKeys.WALLET, value: wallet },
+      };
+    })
+  );
+
+export default [testPageEpic, createWalletEpic];